Resolves issue #14.
If the rawY value is less than zero then the position one list item is below the top of the ScrollView. In this case the expanded quick return view should only slide up as far as the top of the initial list item and the view state should return to the initial STATE_ONSCREEN value.